% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_shaded_quantiles}
\alias{add_shaded_quantiles}
\title{Add a shaded region showing the min/max quantiles}
\usage{
add_shaded_quantiles(
  bayesian_summary,
  col = grDevices::adjustcolor("grey", alpha.f = 0.5),
  ...
)
}
\arguments{
\item{bayesian_summary}{A list-like object of class
\code{bd_bayesian_summary} with information on the quantiles of the density
function and growth rate (see \code{summarize_bayesian_inference}).}

\item{col}{The color of the shaded region (default:
\code{grDevices::adjustcolor("grey",alpha.f = 0.5)}.}

\item{...}{Additional parameters to pass to \code{graphics::polygon}.}
}
\description{
Add a shaded region showing the min/max quantiles (usually +/-2.5\%) to a
plot. The input, bayesian_summary, is the result of a call to
\code{summarize_bayesian_inference}, which is a list-like object of class
bd_bayesian_summary with information on the quantiles of the density function
and growth rate. Typically, the min/max quantiles are 2.5\% and 97.5\%, but
this can be changed via the input lev in the call to
\code{summarize_bayesian_inference}.
}
\seealso{
\code{\link[=summarize_bayesian_inference]{summarize_bayesian_inference()}}
}
