% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_analyze_soln.R
\name{bd_analyze_soln}
\alias{bd_analyze_soln}
\title{Analyze the result of a call to \code{bd_do_inference}}
\usage{
bd_analyze_soln(soln, tau = NA, th_sim = NA, lev = 0.025,
  rateProp = NA, doSummary = T)
}
\arguments{
\item{soln}{The solution, a list-like object of class bd_soln (see \code{bd_do_inference})}

\item{tau}{(optional) The calendar dates at which to evaluate densities.
If tau is not input, tau is built from the hyperparameters.}

\item{th_sim}{(optional) The known parameters used to create simulation data}

\item{lev}{(default: 0.025) The level to use for the quantile bands}

\item{rateProp}{(optional) The cumulative density needed to define rate growth bands}

\item{doSummary}{\link{Default TRUE} Whether to summarize each sample by calling bd_summarize_sample}
}
\value{
A list with information on the quantiles of the density function and growth rate (and sample summaries)
}
\description{
\code{bd_do_inference} calls Stan to do Bayesian inference by generating
a sample of parameters from the posterior of theta (or th).
This function analyzes the result of that inference.
In particular, it calculates the quantiles of the density function and growth rate.
}
\details{
\code{soln} is the result of a call to \code{bd_do_inference}.
It contains both the resulting samples and the parameters used in the inference,
such as the hyperparameters (see \code{bd_do_inference} for further details).
The primary thing \code{bd_analyze_soln} does is calculate quantiles of both
the parameterized density and growth rate. For example,
for a calendar date y each sample yields a density and growth rate.
The quantile is the value of the density or growth rate such that
a given proportion of samples are smaller than that value.
The probabilities used to calculate these quantiles are \code{probs = c(lev, 0.5, 1-lev)},
where \code{lev} is the level (0.025 by default, so that 95% of the observations
lie between the first and last quantile bands).

In addition, \code{bd_analyze_soln} identifies calendar dates for which
the growth rate quantiles defined by \code{lev} and \code{1 - lev} do not contain zero.
This indicates significant positive or negative growth for the density curve.
The output vector \code{growthState} codes calendar dates by growth state as 'negative',
'zero', and 'positive'. For the Gaussian mixture parameterization of the density,
the rate is not typically meaningful near the calendar date boundaries where it
increases linearly as the calendar date goes to positive or negative infinity.
The parameter \code{rateProp} provides control on how calendar dates are classified by
growth rate near these boundaries. In particular, the calendar dates with a cumulative
density (50% quantile) below \code{rateProp} (for the lower boundary) or above \code{1 - rateProp}
(for the upper boundary) are classified as 'missing' in \code{growthState}.
By default, \code{rateProp} is NA and no calendar dates are classified as missing.

By default, a summary is done for each sample by calling
bd_summarize_sample. This is not done of doSummary is FALSE
}
