% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_calc_half_life_from_peak.R
\name{bd_calc_half_life_from_peak}
\alias{bd_calc_half_life_from_peak}
\title{For each sample, calculate the time it takes for the density to decrease by half from the peak}
\usage{
bd_calc_half_life_from_peak(soln, propChange = 0.5, anal = NA,
  peakRange = NA)
}
\arguments{
\item{soln}{The result of a call to bd_do_inference}

\item{propChange}{(Default 0.5) The relative decrease in density to use for the duration calculation}

\item{anal}{(Optional) The result of a call to bd_analyze_soln. If not provided, it is calculated}

\item{(default}{taumin to taumax) peakRange can be given so that the peak density used is on the range peakRange}
}
\value{
A vector of "half-lives" (proportional change set by propChange)
}
\description{
For each sample, calculate the time it takes for the density to decrease by half from the peak
}
\details{
For each sample, calculate the time it takes for the density to decrease by
half from the peak. Optionally, a different proportion can be used than the
default propChange = 0.5. For example, with propChange = 0.1 the time it
takes for the density to decrease by 10% is used. If the relative density is
not reached, the half life for the sample is set to NA. If there is no
interior peak in the range peakRange, which is taumin to taumax by default,
the half life is set to NA.
}
