% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_calc_integration_weights.R
\name{bd_calc_integration_weights}
\alias{bd_calc_integration_weights}
\title{Calculate integration weights assuming the midpoint rule}
\usage{
bd_calc_integration_weights(y)
}
\arguments{
\item{y}{A vector of locations where the function is sampled, possibly irregular}
}
\value{
A vector of integration weights the same length as y
}
\description{
y is a vector of locations where a function to be integrated is
evaluated (the function values, f, are not input). Let
y_g be the locations of integration and f_g the corresponding
function values for g=1, 2, ... G. Assuming trapezoidal
integration at the midpoints between elements of y, the weghts
to use for integration are dy_g = (y_(g+1) - y_(g-1)) / 2, where
the conventions y_0 = y_1 and y_(G+1) = y_G are used.
}
