% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_calc_quantiles.R
\name{bd_calc_quantiles}
\alias{bd_calc_quantiles}
\title{Calculate the quantiles for an input matrix X}
\usage{
bd_calc_quantiles(X, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{X}{The matrix for which quantiles are calculated, with dimensions S x G}

\item{probs}{The probability values at which to calculate the quantiles (default: \code{c(0.025, 0.5, 0.975)})}
}
\value{
The quantiles, a matrix with dimension length(probs) x G
}
\description{
The input matrix X has dimensions S x G, where S is the number
of samples and G the number of grid points at which X was
evaluated. Calculate quantiles for each grid point, g = 1,2,..G.
}
