% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_convert_gauss_mix_param_list_to_vect.R
\name{bd_convert_gauss_mix_param_list_to_vect}
\alias{bd_convert_gauss_mix_param_list_to_vect}
\title{Convert a Gaussian mixture parameterization from a list to a vector}
\usage{
bd_convert_gauss_mix_param_list_to_vect(samp)
}
\arguments{
\item{samp}{A list that parameterizes the Gaussian mixture. See
bayDem_samplePrior.R.}
}
\value{
th A vector that parameterizes the Gaussian mixture. See
bayDem_calcGaussMixPdf.R
}
\description{
For the input, parameterized Gaussian mixture convert from a
list (returned by stan) to a vector representation (which is
more efficient for calculations and the form expected by, e.g.,
bayDem_calcGaussMixPdf.R).
}
