% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_convert_gauss_mix_param_vect_to_list.R
\name{bd_convert_gauss_mix_param_vect_to_list}
\alias{bd_convert_gauss_mix_param_vect_to_list}
\title{Convert a Gaussian mixture parameterization from a vector to a list}
\usage{
bd_convert_gauss_mix_param_vect_to_list(th)
}
\arguments{
\item{th}{A vector that parameterizes the Gaussian mixture. See
bayDem_calcGaussMixPdf.R}
}
\value{
samp A list that parameterizes the Gaussian mixture. See  See
}
\description{
For the input, parameterized Gaussian mixture convert from a
vector representation (which is more efficient for calculations
and the form expected by, e.g., bayDem_calcGaussMixPdf.R) list
(returned by stan).
}
