% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_create_gauss_mix.R
\name{bd_create_gauss_mix}
\alias{bd_create_gauss_mix}
\title{Create a possibly truncated Gaussian mixture}
\usage{
bd_create_gauss_mix(x, taumin = -Inf, taumax = Inf)
}
\arguments{
\item{x}{Parameterization \code{\link{data.frame}} for the Gaussian mixture}

\item{taumin}{(Optional) Lower bound for samples}

\item{taumax}{(Optional) Upper bound for samples}
}
\value{
An object of class \code{\link[distr:UnivarMixingDistribution]{distr::UnivarMixingDistribution}}
}
\description{
Create a possibly truncated Gaussian mixture
}
\details{
\code{x} is a \code{\link{data.frame}} that specifies the Gaussian mixture.
Each row represents a gaussian with a mean of \code{mu},  a standard deviation of \code{sig},
and a probability for the mixing component of \code{pi}. Optionally, the samples are
drawn on the truncated interval \code{taumin} to \code{taumax}.
}
