% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_mixture_model_functions.R
\name{bd_create_gauss_mix_mat}
\alias{bd_create_gauss_mix_mat}
\title{Create a list of possibly truncated Gaussian mixtures for a matrix of samples}
\usage{
bd_create_gauss_mix_mat(x, taumin = -Inf, taumax = Inf, progress = TRUE)
}
\arguments{
\item{x}{A matrix of samples with dimensions S x P}

\item{taumin}{(optional) The mininum time-value for truncation}

\item{taumax}{(optional) The maximum time-value for truncation}

\item{progress}{(optional) Whether to show a progress bar}
}
\value{
A list of objects of class \code{\link[distr:UnivarMixingDistribution]{distr::UnivarMixingDistribution}}
}
\description{
\code{x} is a matrix of samples of a possibly truncated Gaussian
mixture with dimensions S x P, where S is the number of samples
and P is the number of parameters.
\code{bd_create_gauss_mix_mat} supports the same optional
as \code{bd_calc_gauss_mix}: taumin / taumax to specify the
boundaries of truncation.
}
