% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_extract_param.R
\name{bd_extract_param}
\alias{bd_extract_param}
\title{Extract the Bayesian samples for a Gaussian mixture model generated by bd_do_inference}
\usage{
bd_extract_param(fit, asList = F)
}
\arguments{
\item{fit}{The fit from stan, of class stanfit}

\item{asList}{(Default False) Whether to return the samples as a matrix or a list}
}
\value{
A matrix or list of samples
}
\description{
The input fit is the result of a call to stan by \code{bd_do_inference}, of
class stanfit. Return a matrix TH with dimensions S x (3*K), where S is the
number of samples (across all chains, and excluding warmup), and K is the
number of mixtures. Optionally, a list of samples is returned by calling
\code{bd_convert_gauss_mix_param_vect_to_list} repeatedly for each sample.
The final column of as.matrix for class stanfit is the log-posterior, which
must be removed.
}
