% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_norm_pdf_mat.R
\name{bd_norm_pdf_mat}
\alias{bd_norm_pdf_mat}
\title{Normalize the rows of an input density matrix}
\usage{
bd_norm_pdf_mat(fMat, dy)
}
\arguments{
\item{fMat}{The input density matrix with dimensions S x G}

\item{dy}{The grid spacing}
}
\value{
The density matrix with normalized rows
}
\description{
The input density matrix fMat has dimensions S x G,
where S is the number of samples and G is the number of grid points
at which fMat was evaluated.
Normalize each row to integrate to 1 using the input grid spacing dy.
}
