% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_phi2y.R
\name{bd_phi2y}
\alias{bd_phi2y}
\title{Find calendar date from fraction modern value given known bounding indices}
\usage{
bd_phi2y(y_curve, phi_curve, phi_known, ii_lo, ii_hi)
}
\arguments{
\item{y_curve}{Calibration curve calendar dates}

\item{phi_curve}{Calibration curve fraction moderns}

\item{phi_known}{Known fraction modern value}

\item{ii_lo}{Lower bounding index}

\item{ii_hi}{Upper bounding index}
}
\value{
The calendar date corresponding to the input phi_known
}
\description{
Find calendar date from fraction modern value given known bounding indices
}
\details{
y_curve and phi_curve give the calendar date and fraction modern of the
radiocarbon calibration curve. It is known that the calendar date lies
lies between y_curve\link{ii_lo} and y_curve\link{ii_hi}. Interpolate to find the
calendar date corresponding to the input phi_known.
}
