% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{bd_plot_known_sim_density}
\alias{bd_plot_known_sim_density}
\title{Plot the known simulation density}
\usage{
bd_plot_known_sim_density(an, add = F, ...)
}
\arguments{
\item{an}{A list-like object generated by bd_analyze_soln with information on the quantiles of the density function and growth rate}

\item{add}{(default: \code{FALSE}) Whether to make a new plot or add to the active plot}

\item{...}{Additional parameters to pass to plot / lines}
}
\description{
Plot the known simulation density
}
\details{
The input, an, is the result of a call to bd_analyze_soln. It is a
list-like object  with information on the quantiles of the density function
and growth rate. By default, a new plot is made, but if add = TRUE the curve
is added to the active plot.
}
