% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_sample_trunc_exp.R
\name{bd_sample_trunc_exp}
\alias{bd_sample_trunc_exp}
\title{Sample from a truncated exponential distribution}
\usage{
bd_sample_trunc_exp(N, r0, ymin, ymax)
}
\arguments{
\item{N}{The number of samples to make}

\item{r0}{The rate parameter}

\item{ymin}{The lower limit for truncation}

\item{ymax}{The upper limit for truncation}
}
\value{
A vector of N samples between ymin and ymax
}
\description{
Make N samples from a truncated exponential distribution with the density function r0\emph{exp(r0}y). The rate r0 can be positive (growth) or negative (decay). The density function is normalized to integrate to 1 on the interval ymin to ymax.
}
\author{
Michael Holton Price \href{mailto:MichaelHoltonPrice@gmail.com}{MichaelHoltonPrice@gmail.com}
}
