% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability_functions.R
\name{calc_calib_curve_equif_dates}
\alias{calc_calib_curve_equif_dates}
\title{Calculate equifinal dates for each point in the calibration curve}
\usage{
calc_calib_curve_equif_dates(calibDf)
}
\arguments{
\item{calibDf}{The calibration data frame, with columns yearBP, uncalYearBP, and uncalYearBPError}
}
\value{
A list of equifinality information for each point in the calibration curve (see details)
}
\description{
Calculate equifinal dates for each point in the calibration curve
}
\details{
The input calibration data frame has three columns: yearBP, uncalYearBP, and
uncalYearBPError. For each date in the column yearBP, determine all other
dates with the same fraction modern. These dates are equifinal since, in the
absence of additional information, there is no way to determine which year
a sample came from. Typically the actual equifinal date lies between two
observations in yearBP, so linear interpolation is used to estimate the
decimal year. A list of length nrow(calibDf) is returned with, for each
point in the calibration curve, the following information:

indBase -- The row index in calibDf of the point
tauBase -- The calendar date (AD) of the point
indEqui -- The row index/indices in calibDf of the equifinal points
tauEqui -- The calendar date(s) (AD) of the equifinal points
}
