% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{calc_tau_range}
\alias{calc_tau_range}
\title{Calculate a calendar date range that spans the input radiocarbon measurements}
\usage{
calc_tau_range(rc_meas, calibration_curve = "intcal20", dtau = 1)
}
\arguments{
\item{rc_meas}{The radiocarbon measurements (see import_rc_data for the
expected format)}

\item{calibration_curve}{The name of the calibration curve to use (default:
intcal20)}

\item{dtau}{An integer to round to in extending the calendar range on either
end (default: 1, which has no effect)}
}
\value{
A list containing the minimum and maximum calendar dates, tau_min and
tau_max
}
\description{
To determine the calendar dates, first Bchron::BchronCalibrate is called to
calibrate the individual dates, which yields a date range for each. The
maximum range across dates is identified (BchronCalibrate adopts a spacing of
one year), then (if necessary) the range is slighty extended to multiples of
dtau. If intcal20 is being used, dtau should likely be 5, since that is the
spacing at which the calibration curve is specified. If dtau is not an
integer, it is ignored and a warning is thrown. If dtau is 1 (the default),
no rounding is done since that is already the spacing returned by Bchron.
}
\seealso{
\itemize{
\item \code{\link[=set_sim]{set_sim()}} for an overview of the "standard pipeline"
\item \code{\link[=import_rc_data]{import_rc_data()}} for the format of \code{rc_meas}'
}
}
