% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiocarbon_functions.R
\name{calc_trapez_weights}
\alias{calc_trapez_weights}
\title{Calculate integration weights assuming the midpoint rule}
\usage{
calc_trapez_weights(tau)
}
\arguments{
\item{tau}{A vector of locations where the function is sampled, possibly irregularly}
}
\value{
A vector of integration weights the same length as tau
}
\description{
tau is a vector of locations where a function to be integrated
is evaluated (the function values, f, are not input). Let
tau_g be the locations of integration and f_g the corresponding
function values for g=1, 2, ... G. Assuming trapezoidal
integration at the midpoints between elements of tau, the
weights to use for integration are
dtau_g = (tau_(g+1) - tau_(g-1)) / 2, where the conventions
tau_0 = tau_1 and tau_(G+1) = tau_G are used.
}
