% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_lik_fitting_functions.R
\name{calc_trunc_gauss_mix_neg_log_lik}
\alias{calc_trunc_gauss_mix_neg_log_lik}
\title{Calculate the negative log-likelihood of a truncuated Gaussian mixture fit}
\usage{
calc_trunc_gauss_mix_neg_log_lik(th_reduced, M, tau, s_min = 0)
}
\arguments{
\item{th_reduced}{The paramater vector without the first mixture weight}

\item{M}{The measurement matrix}

\item{tau}{The sampling grid}

\item{s_min}{A minimum value for the mixture standard deviations
(default: 0)}
}
\value{
The negative log-likelihood
}
\description{
The input parameter vector is the parameter vector without the first mixture
weight. The minimum and maximum calendar dates are assumed to be the minumum
and maximum values of the sampling grid, tau.
}
