% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_inference_functions.R
\name{do_bayesian_inference}
\alias{do_bayesian_inference}
\title{Sample from the posterior of a density model for a set of radiocarbon
measurements}
\usage{
do_bayesian_inference(
  rc_meas,
  density_model,
  hp,
  calib_df,
  th0 = NA,
  init_seed = NA,
  stan_seed = NA,
  control = list()
)
}
\arguments{
\item{control}{Control parameters (see details)}

\item{prob}{List with the fields \code{phi_m} (vector of fraction modern values
for the radiocarbon measurements, \code{sig_m} (measurement errors for phi_m),
\code{hp} (list of hyperparameters), and \code{calibDf}.}
}
\value{
bayesian_soln, a list with three fields: prob (the input), fit (the result of
the call to stan), and control (the control parameters used)
}
\description{
This is the core function that implements the Bayesian
inference. Currently, the only supported density model is a truncated
Gaussian mixture.
}
\details{
The input consists of the radiocarbon measurements (rc_meas), the
density model (density_model), hyperparameters that specify the prior
distribution as well as the spacing of the sampling grid (hp), and optional
control variables. stan is called via the rstan package to sample from the
posterior. The output is the variable bayesian_soln, which is a list
TODO: continue editing documentation from here once edits are complete
which is a list with the fields prob (the input) and fit (the result of the
stan fit).

prob can also have an optional field control that specifies the
following control parameters for the Bayesian inference (defaults in
parentheses):

num_chains      -- Number of chains (4)
samps_per_chain -- Number of samples per chain (2000)
warmup          -- Number of warmup samples(default is samps_per_chain/2)
init_seed        -- An optional random number seed for initializing the
chains
stan_seed        -- An optional random number seed for the call to Stan
init_list        -- The initializations for each chain. The default is to
set this using a mixture fit of the summed density
}
