% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{draw_rc_meas_using_date}
\alias{draw_rc_meas_using_date}
\title{Simulate radiocarbon measurements from input calendar dates}
\usage{
draw_rc_meas_using_date(t_e, calib_df, error_spec, is_AD = F)
}
\arguments{
\item{t_e}{A vector of event dates (calBP by default)}

\item{calib_df}{The calibration dataframe, with columns year_BP,
uncal_year_BP, and uncal_year_BP_error}

\item{error_spec}{A list with entries type and parameter values specifying
how to add uncertainty to the simulated measurement. Currently, only type =
"unif_fm" is supported, for which the parameters min and max must be in
error_spec.}

\item{is_AD}{Whether t_e is cal_BP or AD (optional; default: \code{FALSE})}
}
\description{
Given input calendar dates t_e, simulate the radiocarbon measurement process
to yield a fraction modern and uncertainty. error_spec is a list with named
type and additional fields specifying the parameters. Currently only type
unif_fm (for uniform fraction modern) is supported, which makes a uniform
draw on the interval min to max for the uncertainty of the fraction modern
value, where min and max are fields in error_spec giving the boundaries of
the uniform draw. By default, t_e is assumed to be calBP, but can be AD (more
more precisely, 1950 - BP value) if is_AD is True.
}
