% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_inference_functions.R
\name{extract_param}
\alias{extract_param}
\title{Extract the Bayesian samples for a Gaussian mixture model generated by
\code{sample_theta}}
\usage{
extract_param(fit)
}
\arguments{
\item{fit}{The fit from stan, of class stanfit}
}
\value{
A matrix of samples with dimensions S by (3*K), where S is the number
of non-warmup samples
}
\description{
The input fit is the result of a call to stan by
\code{sample_theta}, of class stanfit. Return a matrix TH with
dimensions S x (3*K), where S is the number of samples (across all chains,
and excluding warmup), and K is the number of mixtures.
}
