% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{get_aic_vect}
\alias{get_aic_vect}
\title{Extract the vector of Akaike information criterion (AIC) values from a set
of maximum likelihood fits created by do_max_lik_fits}
\usage{
get_aic_vect(max_lik_fits)
}
\arguments{
\item{max_lik_fits}{The list of maximum likelihood fits (see
do_max_lik_fits).}
}
\value{
A vector of AIC values the same length as max_lik_fits.
}
\description{
Extract the vector of Akaike information criterion (AIC) values from a set
of maximum likelihood fits created by do_max_lik_fits
}
\seealso{
\code{\link[=do_max_lik_fits]{do_max_lik_fits()}}
}
