% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{get_best_K}
\alias{get_best_K}
\title{For a mixture model, get the best value of K for an input list of Bayesian
solutions created (based on the widely applicable information criterion,
WAIC).}
\usage{
get_best_K(bayesian_solutions)
}
\arguments{
\item{bayesian_solutions}{The list of Bayesian "solutions" (see
do_bayesian_inference).}
}
\value{
The best value of K
}
\description{
Get the best number of mixture components (K) from the Bayesian inference
based on the widely applicable information criterion (WAIC).
}
\seealso{
\code{\link[=do_bayesian_inference]{do_bayesian_inference()}}
}
