% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{get_best_th}
\alias{get_best_th}
\title{Get the best parameter vector from a set of maximum likelihood fits created
by do_max_lik_fits (by the AIC metric)}
\usage{
get_best_th(max_lik_fits)
}
\arguments{
\item{max_lik_fits}{The list of maximum likelihood fits (see
do_max_lik_fits).}
}
\value{
The best value of K
}
\description{
Get the best parameter vector from a set of maximum likelihood fits created
by do_max_lik_fits (by the AIC metric)
}
\seealso{
\code{\link[=do_max_lik_fits]{do_max_lik_fits()}}
}
