% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{intcal13_raw}
\alias{intcal13_raw}
\title{The raw data for constructing the IntCal13 calibration curve}
\format{
An object of class \code{list} of length 4.
}
\source{
http://intcal.qub.ac.uk/intcal13/query/query.php
}
\usage{
intcal13_raw
}
\description{
A \code{base::list} of \code{tibble::tibble} objects containing the data for calculating the IntCal13 Northern Hemisphere
atmospheric radiocarbon calibration curve.
}
\details{
These data are as available from the IntCal webservice
hosted at Queen's University, Belfast. Details on the structure of the data can be found at
\url{http://intcal.qub.ac.uk/intcal13/query/}.
}
\references{
Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE,
Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H,
Hajdas I, Hatté C, Heaton TJ, Hogg AG, Hughen KA, Kaiser KF, Kromer B,
Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM,
van der Plicht J. IntCal13 and MARINE13 radiocarbon age calibration curves 0-50000 years calBP.
\emph{Radiocarbon} 55(4). DOI: \href{https://doi.org/10.2458/azu_js_rc.55.16947}{10.2458/azu_js_rc.55.16947}
}
\keyword{datasets}
