% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{intcal20}
\alias{intcal20}
\title{The IntCal20 calibration curve}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9500 rows and 5 columns.
}
\source{
http://intcal.org/curves/intcal20.14c
}
\usage{
intcal20
}
\description{
A \link[tibble:tibble]{tibble} containing the IntCal20 Northern Hemisphere
atmospheric radiocarbon calibration curve.
}
\details{
\code{intcal20} is a \link[tibble:tbl_df-class]{tibble::tbl_df} object with the following fields:
\itemize{
\item \strong{CAL BP} --- The calibrated year before present
\item \strong{14C age} --- The un-calibrated 14C age of the sample, in years before present
\item \strong{Error} --- The measurement error of the 14C age, in years before present
\item \strong{Delta 14C} --- The normalized delta 14C (Δ14C) measurement for the sample, in per mil (‰)
\item \strong{Sigma} --- The standard deviation (σ) of the Δ14C measurement, in per mil (‰)
}
}
\references{
Reimer, P. J., Austin, W. E. N., Bard, E., Bayliss, A., Blackwell, P. G.,
Bronk Ramsey, C., Butzin, M., Cheng, H., Edwards, R. L., Friedrich, M.,
Grootes, P. M., Guilderson, T. P., Hajdas, I., Heaton, T. J., Hogg, A. G.,
Hughen, K. A., Kromer, B., Manning, S. W., Muscheler, R., Palmer, J. G.,
Pearson, C., van der Plicht, J., Reimer, R. W., Richards, D. A.,
Scott, E. M., Southon, J. R., Turney, C. S. M., Wacker, L., Adolphi, F.,
Büntgen, U., Capano, M., Fahrni, S. M., Fogtmann-Schulz, A., Friedrich, R.,
Köhler, P., Kudsk, S., Miyake, F., Olsen, J., Reinig, F., Sakamoto, M.,
Sookdeo, A. and Talamo, S. (2020) “THE INTCAL20 NORTHERN HEMISPHERE
RADIOCARBON AGE CALIBRATION CURVE (0–55 CAL kBP),” Radiocarbon.
Cambridge University Press, pp. 1–33.
DOI: \href{https://doi.org/10.1017/RDC.2020.41}{10.1017/RDC.2020.41}
}
\keyword{datasets}
