% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{intcal20_raw}
\alias{intcal20_raw}
\title{The raw data for constructing the IntCal20 calibration curve}
\format{
An object of class \code{list} of length 4.
}
\source{
http://intcal.org/JS/JSintcal20/query/query.php
}
\usage{
intcal20_raw
}
\description{
A \code{base::list} of \code{tibble::tibble} objects containing the data for
calculating the IntCal20 Northern Hemisphere
atmospheric radiocarbon calibration curve.
}
\details{
These data are as available from the IntCal webservice
hosted at Queen's University, Belfast.
Details on the structure of the data can be found at
\url{http://intcal.org/JS/JSintcal20/query/}.
}
\references{
Reimer, P. J., Austin, W. E. N., Bard, E., Bayliss, A., Blackwell, P. G.,
Bronk Ramsey, C., Butzin, M., Cheng, H., Edwards, R. L., Friedrich, M.,
Grootes, P. M., Guilderson, T. P., Hajdas, I., Heaton, T. J., Hogg, A. G.,
Hughen, K. A., Kromer, B., Manning, S. W., Muscheler, R., Palmer, J. G.,
Pearson, C., van der Plicht, J., Reimer, R. W., Richards, D. A.,
Scott, E. M., Southon, J. R., Turney, C. S. M., Wacker, L., Adolphi, F.,
Büntgen, U., Capano, M., Fahrni, S. M., Fogtmann-Schulz, A., Friedrich, R.,
Köhler, P., Kudsk, S., Miyake, F., Olsen, J., Reinig, F., Sakamoto, M.,
Sookdeo, A. and Talamo, S. (2020) “THE INTCAL20 NORTHERN HEMISPHERE
RADIOCARBON AGE CALIBRATION CURVE (0–55 CAL kBP),” Radiocarbon.
Cambridge University Press, pp. 1–33.
DOI: \href{https://doi.org/10.1017/RDC.2020.41}{10.1017/RDC.2020.41}
}
\keyword{datasets}
