% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_lik_fitting_functions.R
\name{is_th_reduced_valid}
\alias{is_th_reduced_valid}
\title{A helper function to determine if the reduced form parameter vector is valid}
\usage{
is_th_reduced_valid(th_reduced, sig_min = 0)
}
\arguments{
\item{th_reduced}{The paramater vector without the first mixture weight}

\item{sig_min}{A minimum value for the mixture standard deviations
(default: 0)}
}
\value{
TRUE or FALSE
}
\description{
The paramater vector is invalid if:
}
\details{
(a) Any mixture weight does not lie between 0 and 1
(b) mu is not ordered from small to large values
(c) A standard deviation is lower than sig_min

Neither bounds on the values of the means or the upper values of the standard
deviations are checked, but they are enforced in the optimization in
bd_fit_trunc_gauss_mix.
}
