% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiocarbon_functions.R
\name{load_calib_curve}
\alias{load_calib_curve}
\title{Load Calibration Curve}
\usage{
load_calib_curve(calib_curve)
}
\arguments{
\item{calib_curve}{Name of calibration curve. One of: "intcal20", "marine20",
"shcal20", "intcal13", "marine13", or "shcal13"}
}
\value{
The calibration dataframe, with columns year_BP, uncal_year_BP, and
uncal_year_BP_error
}
\description{
Parse and return the radiocarbon calibration curve stored in data.
}
