% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{marine13}
\alias{marine13}
\title{The Marine13 calibration curve}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4801 rows and 5 columns.}
\source{
http://www.radiocarbon.org/IntCal13/
}
\usage{
marine13
}
\description{
A \link[tibble:tibble]{tibble} containing the Marine13 marine radiocarbon calibration curve.
}
\details{
\code{marine13} is a \link[tibble:tbl_df-class]{tibble::tbl_df} object with the following fields:
\itemize{
\item \strong{CAL BP} --- The calibrated year before present
\item \strong{14C age} --- The un-calibrated 14C age of the sample, in years before present
\item \strong{Error} --- The measurement error of the 14C age, in years before present
\item \strong{Delta 14C} --- The normalized delta 14C (Δ14C) measurement for the sample, in per mil (‰)
\item \strong{Sigma} --- The standard deviation (σ) of the Δ14C measurement, in per mil (‰)
}
}
\references{
Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE,
Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H,
Hajdas I, Hatté C, Heaton TJ, Hogg AG, Hughen KA, Kaiser KF, Kromer B,
Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM,
van der Plicht J. IntCal13 and MARINE13 radiocarbon age calibration curves 0-50000 years calBP.
\emph{Radiocarbon} 55(4). DOI: \href{https://doi.org/10.2458/azu_js_rc.55.16947}{10.2458/azu_js_rc.55.16947}
}
\keyword{datasets}
