% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{marine13_raw}
\alias{marine13_raw}
\title{The raw data for constructing the Marine13 calibration curve}
\format{
An object of class \code{list} of length 5.
}
\source{
http://intcal.qub.ac.uk/marine/query/query.php
}
\usage{
marine13_raw
}
\description{
A \code{base::list} of \code{tibble::tibble} objects containing the data for calculating the
Marine13 marine radiocarbon calibration curve.
}
\details{
These data are as available from the IntCal webservice
hosted at Queen's University, Belfast. Details on the structure of the data can be found at
\url{http://intcal.qub.ac.uk/marine/query/}.
}
\references{
Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE,
Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H,
Hajdas I, Hatté C, Heaton TJ, Hogg AG, Hughen KA, Kaiser KF, Kromer B,
Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM,
van der Plicht J. IntCal13 and MARINE13 radiocarbon age calibration curves 0-50000 years calBP.
\emph{Radiocarbon} 55(4). DOI: \href{https://doi.org/10.2458/azu_js_rc.55.16947}{10.2458/azu_js_rc.55.16947}
}
\keyword{datasets}
