% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{marine20}
\alias{marine20}
\title{The Marine20 calibration curve}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5500 rows and 5 columns.
}
\source{
http://intcal.org/curves/marine20.14c
}
\usage{
marine20
}
\description{
A \link[tibble:tibble]{tibble} containing the Marine20 marine radiocarbon calibration curve.
}
\details{
\code{marine20} is a \link[tibble:tbl_df-class]{tibble::tbl_df} object with the following fields:
\itemize{
\item \strong{CAL BP} --- The calibrated year before present
\item \strong{14C age} --- The un-calibrated 14C age of the sample, in years before present
\item \strong{Error} --- The measurement error of the 14C age, in years before present
\item \strong{Delta 14C} --- The normalized delta 14C (Δ14C) measurement for the sample, in per mil (‰)
\item \strong{Sigma} --- The standard deviation (σ) of the Δ14C measurement, in per mil (‰)
}
}
\references{
Heaton, T. J., Köhler, P., Butzin, M., Bard, E., Reimer, R. W.,
Austin, W. E. N., Bronk Ramsey, C., Grootes, P. M., Hughen, K. A.,
Kromer, B., Reimer, P. J., Adkins, J., Burke, A., Cook, M. S.,
Olsen, J. and Skinner, L. C. (2020) “MARINE20—THE MARINE RADIOCARBON AGE
CALIBRATION CURVE (0–55,000 CAL BP),” Radiocarbon. Cambridge University
Press, pp. 1–42. DOI: \href{https://doi.org/10.1017/RDC.2020.68}{10.1017/RDC.2020.68}
}
\keyword{datasets}
