% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bd_analysis.R
\name{plot.bd_analysis}
\alias{plot.bd_analysis}
\title{Plot the result of a call to bayDem_analyzeSoln}
\usage{
\method{plot}{bd_analysis}(x, plotType = "density", showSPDF = F, showSim = F, ...)
}
\arguments{
\item{x}{A list-like object of class \code{baydem::bd_analysis} with information on the quantiles of the density function and growth rate}

\item{plotType}{(default: 'density') The type of plot to make.
Options are: 'density', 'rate', and 'shadeddensity'.}

\item{showSPDF}{(default: FALSE) Whether to plot the summed probability.
Ignored unless plotType is 'density'.}

\item{showSim}{(default: FALSE) Whether to add the target density or rate to the plot.}

\item{...}{Additional parameters passed on to other functions.}
}
\description{
Plot the result of a call to bayDem_analyzeSoln. Three types of plots can be created: density, rate, and shaded density. The default is density.
}
\details{
an is the result of a call to bayDem_analyzeSoln.
It is a list-like object of class bayDem_analysis with information
on the quantiles of the density function and growth rate.
The optional variables plotType, showSPDF and showSim control the type of
plot that is made and characteristics of that plot.

The default plotType is 'density'.
This plots the density the 50\% density quantile (solid) and
the bounding quantiles given the level variable, lev,
using in bayDem_analyzeSoln (dashed).
Optionally, the summed probability density and known target distribution
from a simulation can be added.

plotType 'rate' yields a plot of the rate (deriveative of density divided by density)
instead of density. This plots the 50\% rate quantile (solid) and the
bounding quantiles given the level variable, lev, used in bayDem_analyzeSoln (dashed).
Optionally, the rate for the known target distribution from a simulation can be added.
Shaded rectangles are added to the plot using an$growthState with the coloring positive / green,
zero / grey, negative / red, and missing / white.

plotType 'shadeddensity' yields a plot of the density with shaded rectangles
indicating the growth state. As with the plotType 'rate',
this is done using the vector an$growthState with the coloring
positive / green, zero / grey, negative / red, and missing / white.
Optionally, the known target distribution from a simulation can be added.
}
