% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_50_percent_quantile}
\alias{plot_50_percent_quantile}
\title{Plot the 50 percent quantile curve}
\usage{
plot_50_percent_quantile(bayesian_summary, add = F, ...)
}
\arguments{
\item{bayesian_summary}{A list-like object of class
\code{bd_bayesian_summary} with information on the quantiles of the density
function and growth rate (see \code{summarize_bayesian_inference}).}

\item{add}{Whether to make a new plot or add to the active plot (default:
FALSE).}

\item{...}{Additional parameters to pass to plot / lines}
}
\description{
The input, bayesian_summary, is the result of a call to
\code{summarize_bayesian_inference}. By default, a new plot is made, but
if \code{add = TRUE} the curve is added to the active plot.
}
\seealso{
\code{\link[=summarize_bayesian_inference]{summarize_bayesian_inference()}}
}
