% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_model_functions.R
\name{sample_trunc_exp}
\alias{sample_trunc_exp}
\title{Sample from a truncated exponential distribution}
\usage{
sample_trunc_exp(N, r0, tau_min, tau_max)
}
\arguments{
\item{N}{The number of samples to make}

\item{r0}{The rate parameter}

\item{tau_min}{The lower limit for truncation}

\item{tau_max}{The upper limit for truncation}
}
\value{
A vector of N samples between tau_min and tau_max
}
\description{
Make N samples from a truncated exponential distribution with the density
function r0\emph{exp(r0}t). The rate r0 can be positive (growth) or negative
(decay). The density function is normalized to integrate to 1 on the interval
tau_min to tau_max.
}
