% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{set_calib_curve}
\alias{set_calib_curve}
\title{Set the radiocarbon calibration curve for an analysis}
\usage{
set_calib_curve(data_dir, analysis_name, calibration_curve)
}
\arguments{
\item{data_dir}{The directory in which to store analysis data}

\item{analysis_name}{A unique name for a given analysis in data_dir}

\item{calibration_curve}{Either the name of a calibration curve or a data
frame specifying the calibration curve (see load_calib_curve format and
choices).}
}
\description{
This is one of a set of helper functions for undertaking a typical analysis
of radiocarbon dates. For details on the overall framework for these helper
function, see set_rc_meas.
}
\seealso{
\itemize{
\item \code{\link[=set_sim]{set_sim()}} for an overview of the "standard pipeline"
\item \code{\link[=load_calib_curve]{load_calib_curve()}} for the format of \code{calib_df}
}
}
