% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{set_density_model}
\alias{set_density_model}
\title{Set the density model (density_model) for an analysis}
\usage{
set_density_model(data_dir, analysis_name, density_model)
}
\arguments{
\item{data_dir}{The directory in which to store analysis data}

\item{analysis_name}{A unique name for a given analysis in data_dir}

\item{density_model}{A list object specifying the density model}
}
\description{
This is one of a set of helper functions for undertaking a typical analysis
of radiocarbon dates. For details on the overall framework for these helper
function, see set_rc_meas.

set_density_model sets the parameteric model to use for the density function
assumed to give rise to the set of radiocarbon dates (rc_meas). The input
variable density_model is a list in which the field "type" specifies the
type of parametric model and additional named fields are type-specific.
Currently, only one type is supported: "trunc_gauss_mix", which stands for
truncated Gaussian mixture. For type = "trunc_gauss_mix", the following
named fields must also be specified in density_model:

tau_min The minimum calendar year for truncation (in AD or, more precisely,
1950 - years BP)
tau_max The maximum calendar year for truncation (in AD or, more precisely,
1950 - years BP)
K       The number of mixture components. K must be an integer greater than
or equal to 2, and can further be a vector of such integers rather
than a single integer (a vector should be specified if model
selection on the number of mixtures will subsequently be done)
and can be a vector
}
\seealso{
\code{\link[=set_sim]{set_sim()}} for an overview of the "standard pipeline"
}
