% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{shcal13}
\alias{shcal13}
\title{The SHCal13 Southern Hemisphere atmospheric radiocarbon calibration curve}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5141 rows and 5 columns.}
\source{
http://www.radiocarbon.org/IntCal13/
}
\usage{
shcal13
}
\description{
A \code{tibble::tibble} containing the SHCal13 Southern Hemisphere atmospheric radiocarbon calibration curve.
}
\details{
\code{shcal13} is a \link[tibble:tbl_df-class]{tibble::tbl_df} object with the following fields:
\itemize{
\item \strong{CAL BP} --- The calibrated year before present
\item \strong{14C age} --- The un-calibrated 14C age of the sample, in years before present
\item \strong{Error} --- The measurement error of the 14C age, in years before present
\item \strong{Delta 14C} --- The normalized delta 14C (Δ14C) measurement for the sample, in per mil (‰)
\item \strong{Sigma} --- The standard deviation (σ) of the Δ14C measurement, in per mil (‰)
}
}
\references{
Alan G Hogg, Quan Hua, Paul G Blackwell, Caitlin E Buck, Thomas P Guilderson,
Timothy J  Heaton, Mu Niu, Jonathan G Palmer, Paula J Reimer, Ron W Reimer,
Christian S M Turney, Susan R H Zimmerman
\emph{Radiocarbon} 55(4). DOI: \href{https://doi.org/10.2458/azu_js_rc.55.16783}{10.2458/azu_js_rc.55.16783}
}
\keyword{datasets}
