% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{shcal13_raw}
\alias{shcal13_raw}
\title{The raw data for constructing the SHCal13 Southern Hemisphere atmospheric calibration curve}
\format{
An object of class \code{list} of length 3.
}
\source{
http://intcal.qub.ac.uk/shcal13/query/query.php
}
\usage{
shcal13_raw
}
\description{
A \code{base::list} of \code{tibble::tibble} objects containing the data for calculating the
SHCal13 Southern Hemisphere atmospheric radiocarbon calibration curve.
}
\details{
These data are as available from the IntCal webservice
hosted at Queen's University, Belfast. Details on the structure of the data can be found at
\url{http://intcal.qub.ac.uk/shcal13/query/}.
}
\references{
Alan G Hogg, Quan Hua, Paul G Blackwell, Caitlin E Buck, Thomas P Guilderson,
Timothy J  Heaton, Mu Niu, Jonathan G Palmer, Paula J Reimer, Ron W Reimer,
Christian S M Turney, Susan R H Zimmerman
\emph{Radiocarbon} 55(4). DOI: \href{https://doi.org/10.2458/azu_js_rc.55.16783}{10.2458/azu_js_rc.55.16783}
}
\keyword{datasets}
