% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_data.R
\docType{data}
\name{shcal20_raw}
\alias{shcal20_raw}
\title{The raw data for constructing the SHCal20 Southern Hemisphere atmospheric calibration curve}
\format{
An object of class \code{list} of length 3.
}
\source{
http://intcal.org/JS/JSshcal20/query/query.php
}
\usage{
shcal20_raw
}
\description{
A \code{base::list} of \code{tibble::tibble} objects containing the data for calculating the
SHCal20 Southern Hemisphere atmospheric radiocarbon calibration curve.
}
\details{
These data are as available from the IntCal webservice
hosted at Queen's University, Belfast. Details on the structure of the data can be found at
\url{http://intcal.org/JS/JSshcal20/query/}.
}
\references{
Hogg, A. G., Heaton, T. J., Hua, Q., Palmer, J. G., Turney, C. S. M.,
Southon, J., Bayliss, A., Blackwell, P. G., Boswijk, G., Bronk Ramsey, C.,
Pearson, C., Petchey, F., Reimer, P., Reimer, R. and Wacker, L. (2020)
“SHCal20 SOUTHERN HEMISPHERE CALIBRATION, 0–55,000 YEARS CAL BP,”
Radiocarbon. Cambridge University Press, pp. 1–20.
DOI: \href{https://doi.org/10.1017/RDC.2020.59}{10.1017/RDC.2020.59}
}
\keyword{datasets}
