% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_max_lik_fitting.R
\name{temper_trunc_gauss_mix}
\alias{temper_trunc_gauss_mix}
\title{Use parallel tempering to fit a truncated Gaussian mixture to radiocarbon
data}
\usage{
temper_trunc_gauss_mix(
  K,
  phi_m,
  sig_m,
  tau_min,
  tau_max,
  dtau,
  calib_df,
  samps_per_cyc = 10,
  num_cyc = 1000,
  num_cores = NA
)
}
\arguments{
\item{K}{The number of mixture components to use for the fit}

\item{phi_m}{A vector of fraction moderns}

\item{sig_m}{A vector of standard deviations for phi_m}

\item{tau_min}{The minimum calendar date (AD) for the sampling grid}

\item{tau_max}{The maximum calendar date (AD) for the sampling grid}

\item{dtau}{The spacing of the sampling grid}

\item{calib_df}{Calibration curve (see bd_load_calib_curve)}

\item{samps_per_cyc}{Number of samples per cycle for the tempering
(default:10)}

\item{num_cyc}{Number of cycles for the tempering (default:10)}

\item{num_cores}{The number of cores to use for parallelization (default: NA,
do not parallelize)}
}
\value{
A list containing the best fit parameter vector (th), the
corresponding value of the negative log-likelihood (neg_log_lik), tau (the
sampling grid), and f (the density of the best fit paramater vector
evaluated at the points in the sampling grid)
}
\description{
The radiocarbon data are specified by phi_m, the vector of fraction modern
values, and sig_m, the vector of associated standard deviations (both indexed
by i). The likelihood involves an integral over calendar dates that is
calculated using a Riemann integral at the points in the "sampling grid",
tau, which ranges from tau_min to tau_max and has a spacing of dtau (tau is
indexed by the variable g).
}
\details{
Constraints on parameters: The parameter vector, theta (th), has the ordering
th = (pi_1,...,pi_K,mu_1,..mu_K,s_1,...s_K), where K is the number of
mixture components and pi_k / mu_k / s_k are, respectively, the weighting /
mean / standard deviation of the k-th mixture. Optimization is actually done
using a "reduced" vector from which pi_1 has been removed, th = (pi_2,...),
since the mixture weights must add to 1. Constraints are placed on each
variable in the parameter vector. All the weightings must lie between 0 and
1, including pi_1 (even though it is not in th_reduced). The means must lie
between tau_min and tau_max. Since a numerical integral with a defined
spacing is used to calculate the likelihood, the standard deviations of the
mixture components cannot be too small, and are constrained to be greater
than 10*dtau; they are also constrained to be less than tau_max-tau_min. In
summary:\preformatted{  0 < pi_k < 1
}

tau_min < mu_k < tau_max
10*dtau < s_k  < tau_max-tau_min.

The parallel tempering algorithm in the enneal R package is used to do the
optimization. This ensures that the space of possible curves is well-explored
and possible local optima are avoided.
}
