% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{vis_calib_curve}
\alias{vis_calib_curve}
\title{Visualize the calibration curve with equifinal and non-equifinal time spans}
\usage{
vis_calib_curve(
  tau_min,
  tau_max,
  calib_df,
  invert_col = "gray90",
  point_col = "black",
  point_pch = 19,
  ...
)
}
\arguments{
\item{tau_min}{The minimum calendar date for plotting (AD)}

\item{tau_max}{The maximum calendar date for plotting (AD)}

\item{calib_df}{The calibration data frame, with columns year_BP,
uncal_year_BP, and uncal_year_BP_error}

\item{invert_col}{The color for shading invertible regions
(default: \code{gray90})}

\item{point_col}{The color for calibration curve points (default: \code{black})}

\item{point_pch}{The symbol for calibration curve points (default: \code{19})}

\item{...}{Additional inputs to plots}
}
\description{
Vizualize the input calibration curve, calib_df, on the interval tau_min to
tau_max. This involves plotting the curve itself and shading invertible
(non-equifinal) and non-invertible (equifinal) regions.
}
\seealso{
\code{\link[=load_calib_curve]{load_calib_curve()}} for the format of \code{calib_df}
}
