% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability_functions.R
\name{assess_calib_curve_equif}
\alias{assess_calib_curve_equif}
\title{Assess the equifinality or time spans in the input radiocarbon calibration curve}
\usage{
assess_calib_curve_equif(calibDf, equiList = NA)
}
\arguments{
\item{calibDf}{The calibration data frame, with columns yearBP, uncalYearBP, and uncalYearBPError}

\item{equiList}{(Optional) The result of a call to calc_calib_curve_equif_dates}
}
\value{
A list of with named variables canInvert and invSpanlist (see details)
}
\description{
Assess the equifinality or time spans in the input radiocarbon calibration curve
}
\details{
The input calibration data frame has three columns: yearBP, uncalYearBP, and
uncalYearBPError. A time span is equifinal if, for each date in the span,
there is at least one other date in the radiocarbon calibration curve with
the same fraction modern value. Conversely, a time span is not equifinal if
this is not true (all dates in the time span correspond to a unique fraction
modern value). Thus, the calibration curve is divided into alternating
equifinal and non-equifinal spans. This function identifies these regions
for the input calibration data frame, calibDf.

Although calibDf uses year BP, all calculations and returned data use AD.

A list is returned with two named variables: canInvert and invSpanList (inv =
invert, which is conceptually identical to non-equifinal).

canInvert is a boolean that indicates whether each entry (row) in the
calibration data frame, calibDf, is inside or outside an invertible region.

invSpanList is a list summarizing information about all the invertible (non-
equifinal) time spans. It contains the following named variables:

ind       -- Indices (rows) of calibDf following within the invertible time
span
tau_left  -- Calendar date (AD) of the left (earlier) boundary of the
invertible time span
phi_left  -- Fraction modern value of the left (earlier) boundary of the
invertible time span
tau_right -- Calendar date (AD) of the right (later) boundary of the
invertible time span
phi_right -- Fraction modern value of the right (later) boundary of the
invertible time span
ii_prev   -- The index (row) of the calibDf earlier than the invertible
region with closest fraction modern value
ii_next   -- The index (row) of the calibDf later than the invertible
}
