% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_calc_gauss_mix_pdf_mat.R
\name{bd_calc_gauss_mix_pdf_mat}
\alias{bd_calc_gauss_mix_pdf_mat}
\title{Calculate the dnesity of a possibly truncated Gaussian mixture for a matrix of samples}
\usage{
bd_calc_gauss_mix_pdf_mat(TH, y, ymin = NA, ymax = NA,
  type = "density")
}
\arguments{
\item{TH}{A matrix of samples with dimensions S x P}

\item{y}{A vector of points for the density calculation with length G}

\item{ymin}{(optional) The mininum y-value for truncation}

\item{ymax}{(optional) The maximum y-value for truncation}

\item{type}{(optional) The type of calculation: density (default), cumulative, derivative, or rate}
}
\value{
The output matrix with dimensions S x G
}
\description{
TH is a matrix of samples of a possibly truncated Gaussian
mixture with dimensions S x P, where S is the number of samples
and P is the number of parameters. Repeatedly call
\code{bd_calc_gauss_mix_pdf} to calculate the density function
for each sample at the points in the vector y, which has length
G. The output density matrix, fMat, has dimensions S x G.
\code{bd_calc_gauss_mix_pdf_mat} supports the same optional
as \code{bd_calc_gauss_mix_pdf}: ymin / ymax to specify the
boundaries of truncation and type to set whether the density,
cumuluative distribution, derivative, or rate is calculated.
}
