% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_draw_rc_meas_using_date.R
\name{bd_draw_rc_meas_using_date}
\alias{bd_draw_rc_meas_using_date}
\title{Simulate a radiocabron measurement from input calendar dates}
\usage{
bd_draw_rc_meas_using_date(y_e, calibDf, errorSpec, isAD = F)
}
\description{
Given an input calendar dates, simulate the radiocarbon
measurement process to yield a fraction modern and uncertainty.
errorSpec is a list with named components min and max. The
uncertainty, sig_m, is drawn uniformily from this range.
}
