% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_extract_param.R
\name{bd_extract_param}
\alias{bd_extract_param}
\title{Extract the Bayesian samples for a Gaussian mixture model generated by bd_do_inference}
\usage{
bd_extract_param(fit)
}
\arguments{
\item{fit}{The fit from stan, of class stanfit}
}
\value{
A matrix or list of samples
}
\description{
The input fit is the result of a call to stan by \code{bd_do_inference}, of
class stanfit. Return a matrix TH with dimensions S x (3*K), where S is the
number of samples (across all chains, and excluding warmup), and K is the
number of mixtures. The final column of as.matrix for class stanfit is the
log-posterior, which must be removed.
}
