% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiocarbon_functions.R
\name{bd_load_calib_curve}
\alias{bd_load_calib_curve}
\title{Load Calibration Curve}
\usage{
bd_load_calib_curve(calibCurve)
}
\arguments{
\item{calibCurve}{Name of calibration curve}
}
\value{
The calibration dataframe, with columns yearBP, uncalYearBP, and uncalYearBPError
}
\description{
Parse and return the radiocarbon calibration curve stored in data
}
\author{
Michael Holton Price \href{mailto:MichaelHoltonPrice@gmail.com}{MichaelHoltonPrice@gmail.com}
}
