% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_vis_calib_curve.R
\name{bd_vis_calib_curve}
\alias{bd_vis_calib_curve}
\title{Visualize the calibration curve with equifinal and non-equifinal time spans}
\usage{
bd_vis_calib_curve(ymin, ymax, calibDf, invertCol = "gray90",
  pointCol = "black", pointPch = 19, pointCex = 1, ...)
}
\arguments{
\item{ymin}{The minimum calendar date for plotting (AD)}

\item{ymax}{The maximum calendar date for plotting (AD)}

\item{calibDf}{The calibration data frame, with columns yearBP, uncalYearBP, and uncalYearBPError}

\item{invertCol}{\link{Default gray90} The color for shading invertible regions}

\item{pointCol}{\link{Default black} The color for calibration curve points}

\item{...}{Additional inputs to plots}
}
\description{
Visualize the calibration curve with equifinal and non-equifinal time spans
}
\details{
Vizualize the input calibration curve, calibDf, on the interval ymin to
ymax. This involves plotting the curve itself and shading invertible
(non-equifinal) and non-invertible (equifinal) regions.
}
