% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{do_bayesian_inference}
\alias{do_bayesian_inference}
\title{Do Bayesian inference given a set of radiocarbon measurements (rc_meas)
and an existing set of maximum likelihood fits (max_lik_fits)}
\usage{
do_bayesian_inference(
  data_dir,
  analysis_name,
  hp,
  stan_seed = NA,
  control = list()
)
}
\arguments{
\item{data_dir}{The directory in which to store analysis data.}

\item{analysis_name}{A unique name for a given analysis in data_dir.}

\item{hp}{Hyperparameters for the priors and to specify the spacing of the
Riemann sum that approximates the integral for the likelihood (see
sample_theta).}

\item{input_seed}{An integer seed that will be used to set the seed for calls
to the maximum likelihood fits.}

\item{Ccontrol}{arguments to pass to the Bayesian inference function (see
sample_theta).}
}
\description{
This is one of a set of helper functions for undertaking a typical analysis
of radiocarbon dates. For details on the overall framework for these helper
function, see set_rc_meas.

To ensure reproducibility, the input_seed can be provided. The behavior and
expectations are identical to the same parameter used in fit_trunc_gauss_mix,
except that if a vector is specified it should be the same length as the
vector K in analysis$density_model$K. That is, input_seed should be either
(1) a single integer or (2) a vector of length analysis$density_model$K.
}
\seealso{
\code{\link[=set_sim]{set_sim()}} for an overview of the "standard pipeline"
}
