% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{do_bayesian_summary}
\alias{do_bayesian_summary}
\title{Summarize the best model sampled by do_bayesian_inference}
\usage{
do_bayesian_summary(data_dir, analysis_name)
}
\arguments{
\item{data_dir}{The directory in which to store analysis data.}

\item{analysis_name}{A unique name for a given analysis in data_dir.}

\item{plot_file_name}{An optional file name for saving the file}
}
\description{
This is one of a set of helper functions for undertaking a typical analysis
of radiocarbon dates. For details on the overall framework for these helper
function, see set_rc_meas.

Call summarize_bayesian_inference for the best model created by
do_bayesian_inference (the previous step in the standard pipeline). Store the
result in the save file.
}
\seealso{
\code{\link[=do_bayesian_inference]{do_bayesian_inference()}}
}
