% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_blank_density_plot}
\alias{make_blank_density_plot}
\title{Prepare a blank plot for showing densities}
\usage{
make_blank_density_plot(
  an,
  xlim = NA,
  ylim = NA,
  xlab = "Calendar Date [AD]",
  ylab = "Density",
  ...
)
}
\arguments{
\item{an}{A list-like object of class \code{baydem::bd_analysis} with
information on the quantiles of the density function and growth rate}

\item{xlim}{Numeric vector of length 2, giving the x coordinate range}

\item{ylim}{Numeric vector of length 2, giving the y coordinate range}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{Additional parameters to pass to plot}
}
\description{
Prepare a blank plot for showing densities
}
\details{
The input, an, is the result of a call to analyze_soln. It is a
list-like object of class bayDem_analysis with information on the quantiles
of the density function and growth rate. Prepare a blank plot for adding
density information, e.g. by calling plot_50_percent_quantile and
add_shaded_quantiles.R. If the axis labels and limits are not specified,
sensible defaults are used.
}
