% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sample_gauss_mix}
\alias{sample_gauss_mix}
\title{Sample from a possibly truncated Gaussian mixture}
\usage{
sample_gauss_mix(N, th, taumin = NA, taumax = NA)
}
\arguments{
\item{N}{Number of samples}

\item{th}{Parameterization vector for the Gaussian mixture}

\item{taumin}{(Optional) Lower bound for samples}

\item{taumax}{(Optional) Upper bound for samples}
}
\value{
N samples from the Gaussian mixture
}
\description{
Sample from a possibly truncated Gaussian mixture
}
\details{
\code{N} is the number of samples to draw and \code{th} specifies the Gaussian
mixture with the ordering (pi_1,...,pi_K,mu_1,...,mu_K,sig_1,...,sig_K),
where \code{K} is the number of mixture components. Optionally, the samples are
drawn on the truncated interval \code{taumin} to \code{taumax}. Because of limitations
in the package \code{distr}, the maximum number of mixture components suppored is
K=4.
}
