% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{set_rc_meas}
\alias{set_rc_meas}
\title{Set the radiocarbon measurements (rc_meas) for an analysis}
\usage{
set_rc_meas(data_dir, analysis_name, rc_meas)
}
\arguments{
\item{data_dir}{The directory in which to store analysis data}

\item{analysis_name}{A unique name for a given analysis in data_dir}

\item{rc_meas}{The radiocarbon measurements to use for this analysis (the
format of rc_meas is as output by import_rc_data)}
}
\description{
This is one of a set of helper functions for undertaking a typical analysis
of radiocarbon dates. As the analysis proceeds, results are stored in a save
file called analysis_name.rds in the folder data_dir. Where results are
non-determistic, random number seeds are set and stored to ensure that, even
if processing is interupted, results are fully reproducible.
}
\details{
Since this is the first helper function that is used in the sequence of
analysis steps, a full description of the set of helper functions is
described here and the documentation in the other functions points here.

The variables data_dir and analysis_name must be specified for each helper
function. data_dir is a directory that can store the data for multiple
analyses. The analysis_name is a unique specified of a given analysis within
data_dir. If a new analysis is initiated by calling set_rc_meas yet a data
file already exists in data_dir for that analysis name, an error is thrown.

The typical set of steps for an analysis is:

(2) import_rc_data
Import the radiocarbon measurements from file (or simulate them)

(2) set_rc_meas
Set the radiocarbon measurements

(3) calc_tau_range
Determine the range of calendar dates that span the radiocarbon
measurements

(4) set_density_model
Set the parametric model to be used for fitting and Bayesian inference.
Currently, only a truncated Gaussian mixture is supported, which requires
setting the calendar date range for truncation (likely using the result
from step (3), thought it can be explicitly set)
}
