% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io_functions.R
\name{set_sim}
\alias{set_sim}
\title{Use the results of a simulation to set the radiocarbon measurements (rc_meas)
for an analysis (and store the settings used to do the simulation)}
\usage{
set_sim(data_dir, analysis_name, sim)
}
\arguments{
\item{data_dir}{The directory in which to store analysis data}

\item{analysis_name}{A unique name for a given analysis in data_dir}

\item{sim}{The simulation object (see simulate_rc_data)}
}
\description{
This is one of a set of helper functions for undertaking a typical analysis
of radiocarbon dates. As the analysis proceeds, results are stored in a save
file called analysis_name.rds in the folder data_dir. Where results are
non-determistic, random number seeds are set and stored to ensure that, even
if processing is interupted, results are fully reproducible.
}
\seealso{
\code{\link[=set_sim]{set_sim()}} for an overview of the "standard pipeline"
}
